refactor(table): 노션 컨벤션 - 파일 구조에 따라 파일 구조 변경(#1)
