--- src/main/java/com/example/petner/breed/BreedRepository.java
@@ -1,12 +1,0 @@
-package com.example.petner.breed;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface BreedRepository extends JpaRepository<Breed, Long> {
-
-    Optional<Breed> findByName(String name);
-}
