--- src/main/java/com/example/petner/chat/Message.java
@@ -1,45 +1,0 @@
-package com.example.petner.chat;
-
-import com.example.petner.member.Member;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "messages")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Message {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "message_id")
-    private Long messageId;
-
-    @Column(name = "content", nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @CreationTimestamp
-    @Column(name = "sent_at", nullable = false)
-    private LocalDateTime sentAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_room_id", nullable = false)
-    private ChatRoom chatRoom;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "sender_id", nullable = false)
-    private Member sender;
-
-    @Builder
-    public Message(String content, ChatRoom chatRoom, Member sender) {
-        this.content = content;
-        this.chatRoom = chatRoom;
-        this.sender = sender;
-    }
-}
