--- src/main/java/com/example/petner/common/Gender.java
@@ -1,5 +1,0 @@
-package com.example.petner.common;
-
-public enum Gender {
-    MALE, FEMALE
-}
