+++ src/main/java/com/example/petner/domain/chat/entity/ChatRoom.java
@@ -1,0 +1,47 @@
+package com.example.petner.domain.chat.entity;
+
+import com.example.petner.domain.dog.entity.Dog;
+import com.example.petner.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "chat_rooms")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ChatRoom {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "chat_room_id")
+    private Long chatRoomId;
+
+    @CreationTimestamp
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "dog_id")
+    private Dog dog;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id1", nullable = false)
+    private Member member1;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id2", nullable = false)
+    private Member member2;
+
+    @Builder
+    public ChatRoom(Dog dog, Member member1, Member member2) {
+        this.dog = dog;
+        this.member1 = member1;
+        this.member2 = member2;
+    }
+}
