+++ src/main/java/com/example/petner/domain/dog/common/AdoptionStatus.java
@@ -1,0 +1,17 @@
+package com.example.petner.domain.dog.common;
+
+public enum AdoptionStatus {
+    입양_가능("입양 가능"),
+    입양_절차_중("입양 절차 중"),
+    입양_완료("입양 완료");
+
+    private final String displayName;
+
+    AdoptionStatus(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+}
