+++ src/main/java/com/example/petner/domain/favorite/entity/Favorite.java
@@ -1,0 +1,42 @@
+package com.example.petner.domain.favorite.entity;
+
+import com.example.petner.domain.dog.entity.Dog;
+import com.example.petner.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "favorites")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Favorite {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "favorite_id")
+    private Long favoriteId;
+
+    @CreationTimestamp
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "dog_id", nullable = false)
+    private Dog dog;
+
+    @Builder
+    public Favorite(Member member, Dog dog) {
+        this.member = member;
+        this.dog = dog;
+    }
+}
