+++ src/main/java/com/example/petner/domain/location/repository/LocationRepository.java
@@ -1,0 +1,13 @@
+package com.example.petner.domain.location.repository;
+
+import com.example.petner.domain.location.entity.Location;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface LocationRepository extends JpaRepository<Location, Long> {
+
+    Optional<Location> findByStateAndDistrict(String state, String district);
+}
