+++ src/main/java/com/example/petner/domain/post/repository/PostRepository.java
@@ -1,0 +1,26 @@
+package com.example.petner.domain.post.repository;
+
+import com.example.petner.domain.member.entity.Member;
+import com.example.petner.domain.post.entity.Post;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+
+    List<Post> findByAuthor(Member author);
+
+    Page<Post> findAllByOrderByCreatedAtDesc(Pageable pageable);
+
+    @Query("SELECT p FROM Post p WHERE p.title LIKE %:keyword% OR p.content LIKE %:keyword% ORDER BY p.createdAt DESC")
+    Page<Post> searchByTitleOrContent(@Param("keyword") String keyword, Pageable pageable);
+
+    @Query("SELECT p FROM Post p ORDER BY p.viewCount DESC")
+    Page<Post> findAllOrderByViewCountDesc(Pageable pageable);
+}
