+++ src/main/java/com/example/petner/global/config/common/Gender.java
@@ -1,0 +1,5 @@
+package com.example.petner.global.config.common;
+
+public enum Gender {
+    MALE, FEMALE
+}
