--- src/main/java/com/example/petner/location/LocationRepository.java
@@ -1,12 +1,0 @@
-package com.example.petner.location;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface LocationRepository extends JpaRepository<Location, Long> {
-
-    Optional<Location> findByStateAndDistrict(String state, String district);
-}
