--- src/main/java/com/example/petner/member/MemberRepository.java
@@ -1,22 +1,0 @@
-package com.example.petner.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByKakaoId(String kakaoId);
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByNickname(String nickname);
-
-    boolean existsByKakaoId(String kakaoId);
-
-    boolean existsByEmail(String email);
-
-    boolean existsByNickname(String nickname);
-}
