--- src/main/java/com/example/petner/post/Post.java
@@ -1,61 +1,0 @@
-package com.example.petner.post;
-
-import com.example.petner.member.Member;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "posts")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Post {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id")
-    private Long postId;
-
-    @Column(name = "title", nullable = false, length = 200)
-    private String title;
-
-    @Column(name = "content", nullable = false, columnDefinition = "TEXT")
-    private String content;
-
-    @Column(name = "view_count", nullable = false)
-    private Integer viewCount = 0;
-
-    @Column(name = "thumb_image_url", length = 500)
-    private String thumbImageUrl;
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updatedAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "author_id", nullable = false)
-    private Member author;
-
-    @Builder
-    public Post(String title, String content, String thumbImageUrl, Member author) {
-        this.title = title;
-        this.content = content;
-        this.thumbImageUrl = thumbImageUrl;
-        this.author = author;
-        this.viewCount = 0;
-    }
-
-    public void increaseViewCount() {
-        this.viewCount++;
-    }
-}
