--- src/main/java/com/example/petner/shelter/ShelterRepository.java
@@ -1,16 +1,0 @@
-package com.example.petner.shelter;
-
-import com.example.petner.location.Location;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface ShelterRepository extends JpaRepository<Shelter, Long> {
-
-    Optional<Shelter> findByName(String name);
-
-    List<Shelter> findByLocation(Location location);
-}
